+++ internal/cmd/main.go
@@ -1,13 +1,11 @@
 package main
 
 import (
-	ctrl "sigs.k8s.io/controller-runtime"
-
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/cmd/rootcmd"
 )
 
 //go:generate go run github.com/kong/kubernetes-ingress-controller/v2/hack/generators/controllers/networking
 
 func main() {
-	rootcmd.Execute(ctrl.SetupSignalHandler())
+	rootcmd.Execute()
 }