+++ internal/cmd/rootcmd/run.go
@@ -1,13 +1,14 @@
 package rootcmd
 
 import (
-	"context"
 	"fmt"
 
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/manager"
 )
 
-func Run(ctx context.Context, c *manager.Config) error {
+func Run(c *manager.Config) error {
+	ctx := SetupSignalHandler(c.TermDelay)
+
 	diag, err := StartDiagnosticsServer(ctx, manager.DiagnosticsPort, c)
 	if err != nil {
 		return fmt.Errorf("failed to start diagnostics server: %w", err)